+++ package.json
@@ -46,6 +46,7 @@
     "eslint": "5.1.0",
     "eslint-config-prettier": "3.1.0",
     "eslint-plugin-react": "7.1.0",
+    "html-webpack-plugin": "3.2.0",
     "npm-run-all": "4.0.1",
     "redux-mock-store": "1.2.2",
     "rimraf": "2.5.4",
@@ -73,6 +74,6 @@
     "private:serve": "babel-node scripts/serve.js",
     "private:compile": "babel app/ --copy-files --out-dir build",
     "private:clean": "rimraf build",
-    "t::start": "webpack-dev-server --hot --host 0.0.0.0 --config=./webpack/webpack.dev.config.js --mode development"
+    "t::start": "webpack-dev-server --hot --host 0.0.0.0 --config=./dev/webpack.dev.config.js --mode development"
   }
 }
