+++ src/renderer/index.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html>
+<html>  
   <head>
     <title>My App</title>
     <!-- <link rel="stylesheet" type="text/css" href="../../node_modules/typeface-roboto/index.css"> -->
