+++ src/renderer/index.js
@@ -18,7 +18,7 @@ const syncHistoryWithStore = (store, history) => {
 
 const initialState = {};
 const routerHistory = createMemoryHistory();
-const store = configureStore(initialState, routerHistory);
+
 
 syncHistoryWithStore(store, routerHistory);
